+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# CURRENT - v1.0.1
+# 2017-04-02 - v1.1.0
 - Update [Jest](https://www.npmjs.com/package/jest) version
+- Support `#times` method to multiply a string
+- Support `#pad` method to left/right pad a given string
 
 # 2017-02-26 - v1.0.0
 - Update test names to more closely match lib names